Create Serverless-KMS-Grants plugin
Serverless-KMS-Grants is a plugin that will hook into the
deploy and remove serverless lifecycle events. During deploy
it will create a KMS grant for the lamdaRole, if one does
not already exist. During remove, it will revoke the KMS
grant associated with the lambdaRole, if one exists.

The plugin can also be called via the command line.